// src/chat/enums/chat.enum.ts
export enum ChatType {
  DIRECT = 'DIRECT',
  GROUP = 'GROUP',
}

export enum ChatRole {
  OWNER = 'OWNER',
  MEMBER = 'MEMBER',
}

export enum MessageType {
  TEXT = 'TEXT',
  MEDIA = 'MEDIA',
  FILE = 'FILE',
  SYSTEM = 'SYSTEM',
  POLL = 'POLL',
}

export enum PollType {
  MULTI = 'MULTI',
  SINGLE = 'SINGLE',
}